Refactoring of code and which class should be responsible. Created QueueWorker project who should processqueue instead of queuemanager, when implemented. Moved logic for communicating with redis to queuerepositories, now the productRepositories will primarily be responsible for fetching changes from PIM, and writing to cache.
To enable the implementation of queueworker I have adjusted the logic so now instead of working with two sets, we are now also added a list, since BLPOP does not work on sets. Now when products are added to the set, they will then also be added to the list.

By enabling the use of BLPOP we can now being setting up the worker.